feat: added Reviews-feature layers(controller, service, storage), refactored EntityStorage interface, added friends and likes existing checks

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exceptions.DuplicateEventException;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.mappers.FilmMapper;
 import ru.yandex.practicum.filmorate.model.Film;
@@ -135,6 +136,20 @@ public class FilmDbStorage implements FilmStorage {
     @Override
     public void addLike(Integer filmId, Integer userId) {
         this.isFilmExists(filmId);
+
+        String checkQuery =
+                "SELECT film_id " +
+                        "FROM films_likes " +
+                        "WHERE user_id = ? " +
+                        "AND film_id = ?;";
+
+        if (!jdbcTemplate
+                .query(checkQuery, (rs, n) -> rs.getInt("film_id"), userId, filmId)
+                .isEmpty()) {
+            throw new DuplicateEventException(String.format("User with id %d already liked film with id %d",
+                    filmId, userId));
+        }
+
         String query = "merge into films_likes(film_id, user_id) " +
                 "values (?, ?)";
         jdbcTemplate.update(query, filmId, userId);
@@ -143,11 +158,27 @@ public class FilmDbStorage implements FilmStorage {
                 "set rate = rate + 1 \n " +
                 "where id = ?";
         jdbcTemplate.update(updateFilmRate, filmId);
+
+        log.info("like for film with id={} added", filmId);
     }
 
     @Override
     public void deleteLike(Integer filmId, Integer userId) {
         this.isFilmExists(filmId);
+
+        String checkQuery =
+                "SELECT film_id " +
+                        "FROM films_likes " +
+                        "WHERE user_id = ? " +
+                        "AND film_id = ?;";
+
+        if (jdbcTemplate
+                .query(checkQuery, (rs, n) -> rs.getInt("film_id"), userId, filmId)
+                .isEmpty()) {
+            throw new NotFoundException("film", String.format("User with id %d already liked film with id %d",
+                    filmId, userId));
+        }
+
         String query = "delete from films_likes where film_id = ? and user_id = ?";
         jdbcTemplate.update(query, filmId, userId);
 
@@ -155,6 +186,8 @@ public class FilmDbStorage implements FilmStorage {
                 "set rate = rate - 1 \n " +
                 "where id = ?";
         jdbcTemplate.update(updateFilmRate, filmId);
+
+        log.info("like for film with id={} deleted", filmId);
     }
 
     @Override
